Move site select
The former position in the upper right corner suggests that the site selection is somehow global. But actually it's not. Only pages and languages are scoped to the current site, while the rest (users, images, tags) are not. This leads to lots frustration and even errors that no user understands.

By moving the site select into the views where it actually effects something, we reduce confusion and misleading behavior.

This also adds a change in when the language tree select is shown. Only if the current site has mulitple languages we show the select. That's removes clutter from the ui. This behavior also applies to the site select.